Display algorithm preferences list in packet
Adapt Config functions to always return a list of preferred algorithms.
If config is nil, it returns [the_must_algorithm]. If Default<algo> is
given, and it is different from the_must_algorithm, it returns
[the_given_algo, the_must_algorithm]. If any list is given as
<algo>Preferences, then the list is sanitized, such that it contains
the_must_algorithm, and every unsupported algorithm is removed.

This applies to Cipher and Hash. For similar code with AEAD,
an Id dictionary needs to be defined (and probably, declare AEADMode as
an interface).